# Use ulong inside uint256 to improve perf