--- a/NBitcoin/Utils.cs
+++ b/NBitcoin/Utils.cs
@@ -10,6 +10,7 @@ using NBitcoin.Protocol;
 using System.Runtime.ExceptionServices;
 using System.Threading.Tasks;
 using NBitcoin.BouncyCastle.Math;
+using System.Runtime.InteropServices;
 #if !NOSOCKET
 using System.Net.Sockets;
 #endif
@@ -793,6 +794,14 @@ namespace NBitcoin
 #endif
 		public static byte[] ToBytes(uint value, bool littleEndian)
 		{
+#if HAS_SPAN
+			if (littleEndian && BitConverter.IsLittleEndian)
+			{
+				var result = new byte[4];
+				MemoryMarshal.Cast<byte, uint>(result)[0] = value;
+				return result;
+			}
+#endif
 			if (littleEndian)
 			{
 				return new byte[]
@@ -815,10 +824,56 @@ namespace NBitcoin
 			}
 		}
 
+		public static byte[] ToBytes(ulong value, bool littleEndian)
+		{
+#if HAS_SPAN
+			if (littleEndian && BitConverter.IsLittleEndian)
+			{
+				var result = new byte[8];
+				MemoryMarshal.Cast<byte, ulong>(result)[0] = value;
+				return result;
+			}
+#endif
+			if (littleEndian)
+			{
+				return new byte[]
+				{
+					(byte)value,
+					(byte)(value >> 8),
+					(byte)(value >> 16),
+					(byte)(value >> 24),
+					(byte)(value >> 32),
+					(byte)(value >> 40),
+					(byte)(value >> 48),
+					(byte)(value >> 56),
+				};
+			}
+			else
+			{
+				return new byte[]
+				{
+					(byte)(value >> 56),
+					(byte)(value >> 48),
+					(byte)(value >> 40),
+					(byte)(value >> 32),
+					(byte)(value >> 24),
+					(byte)(value >> 16),
+					(byte)(value >> 8),
+					(byte)value,
+				};
+			}
+		}
+
 #if HAS_SPAN
 		public static void ToBytes(uint value, bool littleEndian, Span<byte> output)
 		{
-			if(littleEndian)
+			if (littleEndian && BitConverter.IsLittleEndian)
+			{
+				MemoryMarshal.Cast<byte, uint>(output)[0] = value;
+				return;
+			}
+
+			if (littleEndian)
 			{
 				output[0] = (byte)value;
 				output[1] = (byte)(value >> 8);
@@ -833,42 +888,46 @@ namespace NBitcoin
 				output[3] = (byte)value;
 			}
 		}
-#endif
-
-		public static byte[] ToBytes(ulong value, bool littleEndian)
+		public static void ToBytes(ulong value, bool littleEndian, Span<byte> output)
 		{
+			if (littleEndian && BitConverter.IsLittleEndian)
+			{
+				MemoryMarshal.Cast<byte, ulong>(output)[0] = value;
+				return;
+			}
+
 			if (littleEndian)
 			{
-				return new byte[]
-				{
-					(byte)value,
-					(byte)(value >> 8),
-					(byte)(value >> 16),
-					(byte)(value >> 24),
-					(byte)(value >> 32),
-					(byte)(value >> 40),
-					(byte)(value >> 48),
-					(byte)(value >> 56),
-				};
+				output[0] = (byte)value;
+				output[1] = (byte)(value >> 8);
+				output[2] = (byte)(value >> 16);
+				output[3] = (byte)(value >> 24);
+				output[4] = (byte)(value >> 32);
+				output[5] = (byte)(value >> 40);
+				output[6] = (byte)(value >> 48);
+				output[7] = (byte)(value >> 56);
 			}
 			else
 			{
-				return new byte[]
-				{
-					(byte)(value >> 56),
-					(byte)(value >> 48),
-					(byte)(value >> 40),
-					(byte)(value >> 32),
-					(byte)(value >> 24),
-					(byte)(value >> 16),
-					(byte)(value >> 8),
-					(byte)value,
-				};
+				output[0] = (byte)(value >> 56);
+				output[1] = (byte)(value >> 48);
+				output[2] = (byte)(value >> 32);
+				output[3] = (byte)(value >> 24);
+				output[4] = (byte)(value >> 16);
+				output[5] = (byte)(value >> 8);
+				output[6] = (byte)value;
 			}
 		}
+#endif
 
 		public static uint ToUInt32(byte[] value, int index, bool littleEndian)
 		{
+#if HAS_SPAN
+			if (littleEndian && BitConverter.IsLittleEndian)
+			{
+				return MemoryMarshal.Cast<byte, uint>(value.AsSpan().Slice(index))[0];
+			}
+#endif
 			if (littleEndian)
 			{
 				return value[index]
@@ -887,7 +946,11 @@ namespace NBitcoin
 #if HAS_SPAN
 		public static uint ToUInt32(ReadOnlySpan<byte> value, bool littleEndian)
 		{
-			if(littleEndian)
+			if (littleEndian && BitConverter.IsLittleEndian)
+			{
+				return MemoryMarshal.Cast<byte, uint>(value)[0];
+			}
+			if (littleEndian)
 			{
 				return value[0]
 					   + ((uint)value[1] << 8)
@@ -914,8 +977,45 @@ namespace NBitcoin
 		{
 			return ToUInt32(value, 0, littleEndian);
 		}
-		public static ulong ToUInt64(byte[] value, bool littleEndian)
+
+		public static ulong ToUInt64(byte[] value, int offset, bool littleEndian)
 		{
+#if HAS_SPAN
+			if (littleEndian && BitConverter.IsLittleEndian)
+			{
+				return MemoryMarshal.Cast<byte, ulong>(value.AsSpan().Slice(offset))[0];
+			}
+#endif
+			if (littleEndian)
+			{
+				return value[offset + 0]
+					   + ((ulong)value[offset + 1] << 8)
+					   + ((ulong)value[offset + 2] << 16)
+					   + ((ulong)value[offset + 3] << 24)
+					   + ((ulong)value[offset + 4] << 32)
+					   + ((ulong)value[offset + 5] << 40)
+					   + ((ulong)value[offset + 6] << 48)
+					   + ((ulong)value[offset + 7] << 56);
+			}
+			else
+			{
+				return value[offset + 7]
+					+ ((ulong)value[offset + 6] << 8)
+					+ ((ulong)value[offset + 5] << 16)
+					+ ((ulong)value[offset + 4] << 24)
+					+ ((ulong)value[offset + 3] << 32)
+					   + ((ulong)value[offset + 2] << 40)
+					   + ((ulong)value[offset + 1] << 48)
+					   + ((ulong)value[offset + 0] << 56);
+			}
+		}
+#if HAS_SPAN
+		public static ulong ToUInt64(ReadOnlySpan<byte> value, bool littleEndian)
+		{
+			if (littleEndian && BitConverter.IsLittleEndian)
+			{
+				return MemoryMarshal.Cast<byte, ulong>(value)[0];
+			}
 			if (littleEndian)
 			{
 				return value[0]
@@ -938,6 +1038,12 @@ namespace NBitcoin
 					   + ((ulong)value[1] << 48)
 					   + ((ulong)value[0] << 56);
 			}
+		}
+#endif
+
+		public static ulong ToUInt64(byte[] value, bool littleEndian)
+		{
+			return ToUInt64(value, 0, littleEndian);
 		}