Continuation from https://github.com/MetacoSA/NBitcoin/pull/852

Use ulong instead of uint.

Before

|       Method |      Mean |     Error |    StdDev |
|------------- |----------:|----------:|----------:|
|         Read |  9.020 ns | 0.1607 ns | 0.1424 ns |
| WriteToArray | 14.387 ns | 0.1589 ns | 0.1487 ns |
|  WriteToSpan | 14.172 ns | 0.1521 ns | 0.1348 ns |

After 
|       Method |      Mean |     Error |    StdDev |
|------------- |----------:|----------:|----------:|
|         Read |  7.195 ns | 0.1243 ns | 0.1102 ns |
| WriteToArray | 13.537 ns | 0.2288 ns | 0.2140 ns |
|  WriteToSpan | 13.271 ns | 0.1050 ns | 0.0820 ns |