@NicolasDorier 2020-05-20
I don't reallly understand why, but this is still slower than bech32 parsing...